refactor(launcher): break up the main_app monolith into pure modules + page mixins
main_app.py was a ~4460-line class carrying every page builder, realtime
control, hotkeys and settings — a merge-conflict and maintenance hotspot.
First safe, verified pass at splitting it:

Pure, unit-tested modules (no Tk):
  * app_presets.py    — perf presets + latency-line formatting
  * voice_history.py  — VoiceParamHistory undo/redo snapshot stack
  * audio_devices.py  — monitor-device selection heuristics
main_app methods now thin-delegate to these (behaviour unchanged).

Page-builder mixins under launcher/pages/ composed into MainApp:
  * HomePageMixin   (stage band + carousel)
  * ModelsPageMixin (catalog grid + search/sort)

main_app.py 4461 -> 4018 lines; dead imports removed. Import/MRO verified
with python3.12 (this container's 3.11 lacks tkinter; 3.12 has it), full
pytest suite green (120 passed). docs/LAUNCHER_DECOMPOSITION.md records the
verified mixin pattern and the roadmap for the remaining pages (Settings /
Hotkeys deferred to avoid colliding with the collaborator's active edits).
