+++ launcher/pages/__init__.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+"""Page-builder mixins split out of the main_app monolith.
+
+Each mixin holds one page's Tk builders + handlers and is composed into
+MainApp. They share the same instance (self), so cross-page attributes and
+methods resolve at runtime via the composed class.
+"""
+
+from launcher.pages.home_page import HomePageMixin
+from launcher.pages.models_page import ModelsPageMixin
+
+__all__ = ["HomePageMixin", "ModelsPageMixin"]
